requirement.txt
django==1.8
psycopg2==2.6.2
#celery==3.1.19
#django-celery==3.1
numpy==1.12.0
pandas==0.19.2
requests==2.13.0
django-mptt==0.8.7
django-jsonfield==1.0.1
 django-bootstrap3==8.1.0
